Allow "unlimited" channel limits
On startup, the server creates a store with default store limits.
In order to override those default limits, you can specify them
through Options or command line parameters. To allow a metric to
be ignored, the value -1 (or anything negative) is used to indicate
that the given limit should not be checked.
For instance, you could start a server configuring the max number
of messages, but not limiting the size. You would start the server
like this:

nats-streaming-server --max_msgs 10000 --max_bytes -1

+++ server/server_test.go
@@ -1264,7 +1264,7 @@ func TestMaxBytes(t *testing.T) {
 	msgSize := m.Size()
 	sOpts := GetDefaultOptions()
 	sOpts.ID = clusterName
-	sOpts.MaxBytes = uint64(msgSize * 10)
+	sOpts.MaxBytes = int64(msgSize * 10)
 	s := RunServerWithOpts(sOpts, nil)
 	defer s.Shutdown()
 
@@ -1285,7 +1285,7 @@ func TestMaxBytes(t *testing.T) {
 	if err != nil {
 		t.Fatalf("Unexpected error getting state: %v", err)
 	}
-	if b != sOpts.MaxBytes {
+	if b != uint64(sOpts.MaxBytes) {
 		t.Fatalf("Expected msgs size to be %v, got %v", sOpts.MaxBytes, b)
 	}
 }
